Compile command
Added the ability in the compile module to produce a command which will
be used to compile the document, as well as functionality in the Outline
struct to produce a valid string of LaTeX which can be passed to the
command. Currently, the only option is to compile the entire document,
not individual pieces.