Initial Support for Class Options
Added support for class options. These are to be specified in the
command line after the -o argument. Since the program now needs to know
what the document class is, a 'class' line should be specified in the
config file. Currently, the program does not check to make sure that
this is set up properly.